refactor: activity->fragment 전환 intent 대신 finish()로 변경
