--- app/src/main/java/com/example/woomansi/ui/screen/main/Main2Fragment.java
@@ -1,28 +1,0 @@
-package com.example.woomansi.ui.screen.main;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.fragment.app.Fragment;
-
-import com.example.woomansi.R;
-
-public class Main2Fragment extends Fragment {
-
-    public Main2Fragment() {
-        // Required empty public constructor
-    }
-
-    public static Main2Fragment newInstance() {
-        return new Main2Fragment();
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_main2, container, false);
-    }
-}
