+++ app/src/main/java/com/example/woomansi/ui/screen/main/MainActivity.java
@@ -1,16 +1,10 @@
 package com.example.woomansi.ui.screen.main;
 
-import android.content.Intent;
-import androidx.annotation.Nullable;
+import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
-
-import android.os.Bundle;
-
 import com.example.woomansi.R;
-import com.example.woomansi.ui.screen.group.GroupMainFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
-import java.util.Objects;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -28,15 +22,7 @@ public class MainActivity extends AppCompatActivity {
             return true;
         });
 
-        String fragment = getIntent().getStringExtra("전환할 화면");
-
-        if(Objects.equals(fragment, "그룹"))
-        {
-            bottomNav.setSelectedItemId(R.id.item_group);
-        }
-        else {
-            bottomNav.setSelectedItemId(R.id.item_home);
-        }
+        bottomNav.setSelectedItemId(R.id.item_home);
     }
 
     private void changeFragment(int itemId) {
