+++ app/src/main/res/layout/dialog_join_group_fail.xml
@@ -10,13 +10,13 @@
     android:elevation="10dp">
 
     <TextView
-        android:id="@+id/Exit"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:text="그룹을 찾지 못하였습니다.\n다시 입력해주세요!"
-        android:textAppearance="@style/black_bold"
-        android:textSize="20dp" />
+      android:id="@+id/Exit"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:layout_centerInParent="true"
+      android:text="그룹을 찾지 못하였습니다.\n다시 입력해주세요!"
+      android:textAppearance="@style/black_bold"
+      android:textSize="20dp" />
 
     <RelativeLayout
         android:layout_width="match_parent"
